requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
eval_type_backport==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
Markdown==3.8.2
mistralai==1.8.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
